+++ docs/.vitepress/theme/index.ts
@@ -1,13 +1,8 @@
 import DefaultTheme from 'vitepress/theme'
-import { h } from 'vue'
-import StyleSwitcher from './StyleSwitcher.vue'
-import 'lxgw-wenkai-screen-webfont/lxgwwenkaiscreen.css'
+import ThemeLayout from './ThemeLayout.vue'
 import './custom.css'
 
 export default {
   extends: DefaultTheme,
-  Layout: () =>
-    h(DefaultTheme.Layout, null, {
-      'nav-bar-content-after': () => h(StyleSwitcher)
-    })
+  Layout: ThemeLayout
 }